Move the rest of casting.rs
